chore: apply cargo clippy

+++ lib/src/timer.rs
@@ -639,10 +639,7 @@ mod test {
         let mut t = timer();
 
         t.set_timeout_at(Duration::from_millis(100), "a");
-        t.set_timeout_at(
-            Duration::from_millis((100 + TICK * SLOTS as u64) as u64),
-            "b",
-        );
+        t.set_timeout_at(Duration::from_millis((100 + TICK * SLOTS as u64)), "b");
 
         let mut tick = ms_to_tick(&t, 100);
         assert_eq!(Some("a"), t.poll_to(tick));
